requirement.txt
Django==3.2.14
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-import-export==2.8.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-phonenumber-field==6.3.0
django-storages==1.12.3
django-user-agents==0.4.0
django-user-visit==0.5.1
djangorestframework==3.13.1
phonenumbers==8.12.51
Pillow==9.2.0
jsonfield
boto3==1.24.31
youtube-dl
PyPDF2
pafy
pytesseract
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
django-user-agents
pandas
jwt
pyotp
channels==3.0.5
timeago==1.0.15
braintree
nexmo_jwt
Werkzeug==2.1.2
sentry-sdk==1.27.1
django-select2==8.1.2
fuzzywuzzy==0.18.0 
html2text==2020.1.16 
vonage==3.7.1
PyJWT==2.7.0
googletrans==3.0.0
bs4==0.0.1
django-allauth==0.54.0